Moved command executor to a new file. Playing with redirections
