+++ Makefile
@@ -6,7 +6,7 @@
 #    By: limartin <limartin@student.codam.nl>         +#+                      #
 #                                                    +#+                       #
 #    Created: 2020/10/14 11:19:17 by limartin      #+#    #+#                  #
-#    Updated: 2020/10/29 13:37:54 by jsaariko      ########   odam.nl          #
+#    Updated: 2020/10/30 16:07:37 by jsaariko      ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
@@ -49,6 +49,7 @@ SRC = $(SRC_PATH)main.c \
 	$(SRC_PATH)$(PARSE_PATH)comp_links.c \
 	$(SRC_PATH)$(PARSE_PATH)free_and_exit.c \
 	$(SRC_PATH)$(EX_PATH)execute.c \
+	$(SRC_PATH)$(EX_PATH)manage_execution.c \
 	$(SRC_PATH)$(EX_PATH)echo.c \
 	$(SRC_PATH)$(EX_PATH)env.c \
 	$(SRC_PATH)$(EX_PATH)env_utils.c \
