+++ log.log
@@ -1,0 +1,16 @@
+Makefile
+README
+includes
+libft
+log.log
+minishell
+minishell.en.subject.pdf
+srcs
+Makefile
+README
+includes
+libft
+log.log
+minishell
+minishell.en.subject.pdf
+srcs
