+++ srcs/execute/execute.c
@@ -6,15 +6,13 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/22 16:32:46 by jsaariko      #+#    #+#                 */
-/*   Updated: 2020/10/29 18:49:29 by jsaariko      ########   odam.nl         */
+/*   Updated: 2020/10/30 16:08:25 by jsaariko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "execute.h"
 #include "libft.h"
-#include <unistd.h>
-#include "error.h"
-#include <dirent.h>
+
 
 /*
 ** TODO: Build command validators in functions
@@ -33,111 +31,17 @@ t_cmd	get_command(t_icomp *comp)
 	return (NULL);
 }
 
-char	**build_argv(t_icomp *comp)
-{
-	int i = 0;
-	int j = 0;
-	char **try = ft_split(comp->arg, ' ');
-	while(try[i] != NULL)
-	{
-		ft_dprintf(STDIN_FILENO, "%s, %d\n", try[i], i);
-		i++;
-	}
-	char **argv = (char **)e_malloc(i + 1);
-	argv[j] = comp->cmd;//change
-	j++;
-	i = 0;
-	while(try[i] != NULL)
-	{
-		ft_dprintf(STDIN_FILENO, "lol %d %d\n", i, j);
-		argv[j] = try[i];
-		i++;
-		j++;
-	}
-	argv[j] = NULL;
-	//free matrix
-	return (argv);
-}
-
-char	*readdir_test(t_vector *env, t_icomp *comp)
-{
-	t_env *item = vector_get(env, vector_search(env, compare_key, "PATH"));
-	// ft_dprintf(STDIN_FILENO, "%s\n", item->value);
-	char **split = ft_split(item->value, ':');
-	int i;
-	i = 0;
-	char *found;
-	found = NULL;
-	while(split[i] != NULL)
-	{
-		// ft_dprintf(STDIN_FILENO, "path part %d: %s\n", i, split[i]);
-		i++;
-	}
-	// ft_dprintf(STDIN_FILENO, "DONE\n");
-	DIR *dir;
-	int j = 0;
-	while(split[j] != NULL)
-	{
-		dir = opendir(split[j]);
-		if (dir == NULL)
-			break ;
-		struct dirent *entry;
-		entry = readdir(dir);
-		while ((entry = readdir(dir)) != NULL)
-		{
-			if (ft_strncmp(entry->d_name, comp->cmd, ft_strlen(comp->cmd) + 1) == 0)
-			{
-				found = ft_strjoin(split[j], "/");
-				closedir(dir); //TODO: free all the shit
-				return (found);
-			}
-		}
-		closedir(dir);
-		j++;
-	}
-	return (NULL);
-	(void)comp;
-}
-
-void	exec_command(t_vector *env, t_icomp *comp)
-{
-	int pid;
-	char **envp;
-	char **argv;
-
-	char *dir= readdir_test(env, comp);
-	if (dir == NULL)
-		invalid_cmd(comp);
-	char *command = ft_strjoin(dir, comp->cmd);
-	pid = fork();
-	if (pid != 0) //if in parent process
-	{
-		wait(&pid);
-	}
-	else
-	{
-		envp = env_to_envp(env);
-		free_environment(env);
-		argv = build_argv(comp);
-		execve(command, argv,  envp);
-		exit(0);
-	}
-	
-}
-
 void	execute(t_vector *env, t_icomp *comp)
 {
-	t_cmd	f;
 	t_icomp	*tmp;
+	int i;
 
+	i = 0;
 	tmp = comp;
 	while (tmp != NULL)
 	{
-		f = get_command(tmp);
-		if (f == NULL)
-			exec_command(env, comp);
-		else
-			f(env, tmp);
+		exec_command(env, tmp);
 		tmp = tmp->right;
+		i++;
 	}
 }
