+++ srcs/invalid.c
@@ -6,7 +6,7 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/27 10:09:20 by jsaariko      #+#    #+#                 */
-/*   Updated: 2020/10/29 09:37:21 by jsaariko      ########   odam.nl         */
+/*   Updated: 2020/10/30 13:52:02 by jsaariko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
