+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/14 11:59:41 by jsaariko      #+#    #+#                 */
-/*   Updated: 2020/10/29 14:08:10 by jsaariko      ########   odam.nl         */
+/*   Updated: 2020/10/30 13:51:59 by jsaariko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,9 @@ int		main(int ac, char **av, char **envp)
 	(void)ac;
 	(void)av;
 	while (1)
+	{
 		get_input(env);
+	}
 	return (0);
 }
